Implemented basic key extractor
Can currently select a key to record, for 5 seconds the key should be pressed, the audio is then analysed and each key press is sampled and saved as a wav. This will be useful for generating training data for the models. Very primitive method but does the job, parameters can be altered as required (e.g. smaller threshold for a quieter keyboard)